refactor(ondevicemessage): do not cache if device model is not defined
The device table is eventual consitent and in tests
the first time the device model is looked up the entry
might not exist. However previously the fallback value
"default" was cached and used in all subsequent retries.

This changes the caching strategy to only cache "good"
values and removes the "default" model which is not supported
anyway.

Fixes #31