feat(websocket): improve accuracy of active connections
This reduces the TTL in the connections table to 5 minutes.
Clients have the send a message to push the TTL window forward.
In addition, a lambda authorizer is introduced that handles
the authentication based on the fingerprint query string and
makes a context object available for the websocket lambdas
that contains the deviceId and the model.

The shadow fetcher now uses the ttl to reduce the number
of connections it considers active

In addition, the shadow fetcher re-uses the existing connection table