feat(cli): align configure command
This changes the configure command to follow the other acount
related commands and now accepts an account name instead of
a scope.

In addition, the nRFCloud/accounts/<account> with value <account>
SSMParameter has been removed because the list of active accounts
can be determined by looking at which accounts have been configured.