feat(history): implement location trail (#178)
See https://github.com/hello-nrfcloud/web/issues/68

This also changes the the historical data code and separates parts that deal with historical sensor data and the ones that deal with location data. This was necessarily because a new location history request was added that is not handled by the repository.

It also removes the dependency to concrete transformed messages in the backend for the description of requests and responses and uses generic ones.