fix: increase concurrency rate
We expected the fetchDeviceShadow function to be invoked 12 times per minute, as it is triggered by SQS every 5 seconds. However, during the load test we observed that the average invocation rate was only 6 times per minute. After a thorough investigation, we determined that this discrepancy was due to a concurrency limit of 2 on the function. To align with our desired invocation rate, we increase the maxConcurrency to 15.