fix: add retry logic to websocket connection
There is an issue that test is failed due to websocket connection, https://github.com/hello-nrfcloud/backend/actions/runs/6183891652/job/16786519316?pr=259.

There is no log what cause the connection is failed. So, this PR will introduce logging error when the connection is failed and also retry logic to websocket connection.