refactor: use higher-level API for validatedFetch
This updates the validateFetch API and integrates
the fetch configuration in the first parameter.

This way there are not two parameters (that
are separated by the schema parameter) that control
what fetch sends.

It also adds convenience functions to provide
payloads and sends the correct content-type header
only if needed.