refactor: pull out Docker container management
This removes the unneccessary integration of the Docker
image creation and publication with the backend stack.

There was a needless dependency between the repository
and the deployment of the stack (it included it in
its output), while the repo is known already before
the deployment.

This also separates the concern of preparing the containers
from the deployment which allows to prepare containers
individually.

No need to check for containers before deploy
It will happen anyway in in CloudFormation.